ExtractOGMetadata code converted to utils
