+++ src/app/[...slug]/page.tsx
@@ -2,6 +2,7 @@ import RedirectPageBySlugData from "./RedirectPageBySlugData";
 import { Metadata } from "next";
 import linksListModel from "@/models/linksListModel";
 import { connect2MongoDB } from "connect2mongodb";
+import { ExtractOGMetadata } from "@/utils/ExtractOGMetadata";
 
 async function getDestinationURL(alias: string) {
     'use server'
@@ -12,36 +13,12 @@ async function getDestinationURL(alias: string) {
     }
 }
 
-const extractOGMetadata = async (url: string | URL | Request) => {
-    try {
-        const response = await fetch(url);
-        const html = await response.text();
-
-        const ogTitleMatch = html.match(/<meta\s+property="og:title"\s+content="([^"]+)"\s*\/?>/i);
-        const ogImageMatch = html.match(/<meta\s+property="og:image"\s+content="([^"]+)"\s*\/?>/i);
-        const ogDescriptionMatch = html.match(/<meta\s+property="og:description"\s+content="([^"]+)"\s*\/?>/i);
-
-        const ogTitle = ogTitleMatch ? ogTitleMatch[1] : null;
-        const ogImage = ogImageMatch ? ogImageMatch[1] : null;
-        const ogDescription = ogDescriptionMatch ? ogDescriptionMatch[1] : null;
-
-        return { ogTitle, ogImage, ogDescription };
-    } catch (error) {
-        console.error('Error fetching OG metadata:', error);
-        return {
-            ogTitle: 'Shave - Short Your Links',
-            ogDescription: 'Shorten URLs effortlessly for seamless sharing across all platforms with ease.',
-            ogImage: '../opengraph-image.png'
-        };
-    }
-};
-
 export async function generateMetadata({ params }: { params: { slug: [string, string] } }): Promise<Metadata> {
     if (params.slug[0] !== "_next") {
         const url = await getDestinationURL(params.slug[0]);
 
         const fetchData = async () => {
-            const metadata = await extractOGMetadata(url);
+            const metadata = await ExtractOGMetadata(url);
             return metadata;
         };
